Refactor TestHttpCredentialsProvider
